End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -19,7 +20,7 @@ export class MessagesComponent implements OnInit {
 
 
 
-  constructor(private messageService:MessageService) { }
+  constructor(private messageService:MessageService,private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -36,10 +37,17 @@ loadMessages(){
 }
 
 deleteMessage(id:number){
-  this.messageService.deleteMessage(id).subscribe(()=>
-  {
-    this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
-  })
+  this.confirmService.confirm('Confirm Delete Message','This cannot be undone').subscribe(result=>
+    {
+    if(result){
+      this.messageService.deleteMessage(id).subscribe(()=>
+      {
+        this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
+      })
+    }
+    })
+
+
 }
 pageChanged(event:any){
   if(this.pageNumber!==event.page)
